Uniformly use 64 bit software timers.

+++ src/color.cpp
@@ -39,7 +39,7 @@ void Color_FadeSystemPalette(uint8_t* src, uint8_t* dest, int32_t time_limit) {
     uint8_t palette[PALETTE_STRIDE * PALETTE_SIZE];
 
     for (int32_t time_spent{0}, measured_time{1}; time_spent < time_limit; time_spent += measured_time) {
-        const uint32_t time_stamp = timer_get();
+        const uint64_t time_stamp = timer_get();
 
         for (uint32_t j = 0; j < sizeof(palette); ++j) {
             const int32_t d = src[j] - dest[j];
